Switch to Fall2024 datagen, create testing CLI
