+++ .gitignore
@@ -25,3 +25,5 @@ rust-project.json
 # End of https://www.toptal.com/developers/gitignore/api/rust,rust-analyzer
 
 *.DS_store
+
+*.vscode
