+++ datagen/Cargo.toml
@@ -4,6 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+clap = { version = "4.5.1", features = ["derive"] }
+colored = "2.1.0"
 regex = "1.10.2"
 reqwest = { version = "0.11.22", features = ["blocking"] }
 scraper = "0.18.1"
