+++ datagen/data/departments_list.txt
@@ -1,5 +1,7 @@
+AAPS
 AASP
 AAST
+ABRM
 AGNR
 AGST
 AMSC
@@ -17,9 +19,7 @@ ARTH
 ARTT
 ASTR
 BCHM
-BDBA
 BIOE
-BIOI
 BIOL
 BIOM
 BIPH
@@ -27,8 +27,6 @@ BISI
 BMGT
 BMSO
 BSCI
-BSCV
-BSGC
 BSOS
 BSST
 BUAC
@@ -53,6 +51,7 @@ CMLT
 CMSC
 COMM
 CPBE
+CPCV
 CPET
 CPGH
 CPJT
@@ -138,23 +137,20 @@ LBSC
 LEAD
 LGBT
 LING
-MAIT
 MATH
 MEES
 MIEH
 MITH
 MLAW
 MLSC
-MSML
-MSQC
+MOCB
 MUED
 MUSC
 NACS
 NAVY
 NEUR
 NFSC
-NIAP
-NIAV
+NIAS
 PEER
 PERS
 PHIL
@@ -184,8 +180,6 @@ UMEI
 UNIV
 URSP
 USLT
-VIPS
 VMSC
 WEID
 WGSS
-XPER
